# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for ISHC project.",
        "title": "ISHC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Get details of all events",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get all events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new event with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "Event payload",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.EventInfo"
                        }
                    }
                }
            }
        },
        "/events/oldperson/{oldperson_id}": {
            "get": {
                "description": "Get details of events by old person ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get events by old person ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Old person ID",
                        "name": "oldperson_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventInfo"
                            }
                        }
                    }
                }
            }
        },
        "/events/search": {
            "get": {
                "description": "Search events with the given parameters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Search events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event type",
                        "name": "event_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Event date",
                        "name": "event_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Event location",
                        "name": "event_location",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Event description",
                        "name": "event_desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Old person ID",
                        "name": "oldperson_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventInfo"
                            }
                        }
                    }
                }
            }
        },
        "/events/task/{task_id}": {
            "get": {
                "description": "Get details of events by task ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get events by task ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventInfo"
                            }
                        }
                    }
                }
            }
        },
        "/events/type/{type}": {
            "get": {
                "description": "Get details of events by type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "Get events by type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventInfo"
                            }
                        }
                    }
                }
            }
        },
...
        "/oldpersons": {
            "get": {
                "description": "Get details of all old persons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Get all old persons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OldPersonInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new old person with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Create a new old person",
                "parameters": [
                    {
                        "description": "Old Person payload",
                        "name": "oldPerson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OldPersonInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OldPersonInfo"
                        }
                    }
                }
            }
        },
...
        "/oldpersons/{id}": {
            "get": {
                "description": "Get details of an old person by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Get an old person by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Old Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OldPersonInfo"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an old person with the given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Update an old person",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Old Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Old Person payload",
                        "name": "oldPerson",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OldPersonInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OldPersonInfo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an old person by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Delete an old person",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Old Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/oldpersons/{id}/profile_photo": {
            "post": {
                "description": "Set profile photo for an old person by ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oldperson"
                ],
                "summary": "Set old person profile photo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Old Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Profile Photo URL",
                        "name": "profile_photo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "models.CustomTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
        "models.EventInfo": {
            "type": "object",
            "properties": {
                "event_date": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "event_desc": {
                    "type": "string"
                },
                "event_location": {
                    "type": "string"
                },
                "event_type": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "oldperson_id": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.OldPersonInfo": {
            "type": "object",
            "properties": {
                "birthday": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "checkin_date": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "checkout_date": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "created": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "firstguardian_name": {
                    "type": "string"
                },
                "firstguardian_phone": {
                    "type": "string"
                },
                "firstguardian_relationship": {
                    "type": "string"
                },
                "firstguardian_wechat": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "health_state": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_card": {
                    "type": "string"
                },
                "imgset_dir": {
                    "type": "string"
                },
                "isactive": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "remove": {
                    "type": "string"
                },
                "room_number": {
                    "type": "string"
                },
                "updated": {
                    "$ref": "#/definitions/models.CustomTime"
                },
                "updated_by": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}